Use int random numbers

+++ include/kernel/resampling.h
@@ -5,7 +5,7 @@
 #include <thrust/device_vector.h>
 #include <thrust/host_vector.h>
 
-void calc_resampled_indeces(thrust::device_vector<float>& weight_accum, thrust::device_vector<float>& rand_array,
+void calc_resampled_indeces(thrust::device_vector<float>& weight_accum, thrust::device_vector<int>& rand_array,
 	thrust::device_vector<int>& indices);
 
 __global__ void resamplingKernel(Particle* particle_array, Particle* particle_array_next, Particle* birth_particle_array,
